QoL Fixes:
- Move optimization objects into separate classes
- Add logging properly instead of print statements
- Formatting
