+++ databallpy/optimization/tester.py
@@ -1,33 +1,23 @@
-from databallpy.optimization.simulated_annealing import SimulatedAnnealing
-from kloppy import skillcorner
-from databallpy import get_game_from_kloppy
-import matplotlib.patches as mpatches
 
-from databallpy.visualize import (
-    _plot_heatmap_overlay,
-    _plot_single_frame,
-    _plot_velocities,
-    pick_bw_for_contrast,
-    plot_soccer_pitch,
-    plot_tracking_data,
-)
-from matplotlib.colors import to_rgb
+import numpy as np
+from kloppy import skillcorner
+from matplotlib.colors import LinearSegmentedColormap
+
+from databallpy import get_game_from_kloppy
 from databallpy.features.pitch_control import (
     get_pitch_control_single_frame,
     get_team_influence,
 )
-from databallpy.optimization.optimization import (
-    WeightedPitchControlObjective,
+from databallpy.optimization.constraints import TTIConstraint
+from databallpy.optimization.objectives import (
     PressureObjective,
-    TTIConstraint,
-    optimize_tracking_frame,
+    WeightedPitchControlObjective,
 )
-import pandas as pd
-import numpy as np
-from matplotlib.colors import LinearSegmentedColormap
-from matplotlib.legend_handler import HandlerTuple
-from matplotlib.lines import Line2D
-from copy import deepcopy
+from databallpy.optimization.simulated_annealing import SimulatedAnnealing
+from databallpy.visualize import (
+    plot_soccer_pitch,
+    plot_tracking_data,
+)
 
 
 def generate_fig(game, frame_idx, save_path=None):
@@ -60,28 +50,27 @@ def generate_fig(game, frame_idx, save_path=None):
         fig.savefig(save_path)
     return fig, ax
 
-def diff_frames(game_1, frame_idx_1, game_2, frame_idx_2, save_path=None):
-    # frame_1 = game_1.tracking_data[game_1.tracking_data["frame"] == frame_idx_1].iloc[0]
-    # frame_2 = game_2.tracking_data[game_2.tracking_data["frame"] == frame_idx_2].iloc[0]
-    # pitch_control_1 = get_pitch_control_single_frame(
-    #     frame_1,
-    #     game_1.pitch_dimensions,
-    #     n_x_bins=106,
-    #     n_y_bins=68,
-    # )
-    # pitch_control_2 = get_pitch_control_single_frame(
-    #     frame_2,
-    #     game_2.pitch_dimensions,
-    #     n_x_bins=106,
-    #     n_y_bins=68,
-    # )
-    # cmap_red_green = LinearSegmentedColormap.from_list(
-    #     "reds", [(0, 1, 0, 1), (0.5, 0.5, 0, 0), (1, 0, 0, 1)]
-    # )
 
-    fig, ax = plot_soccer_pitch(
-        field_dimen=game_1.pitch_dimensions, pitch_color="white"
+def diff_frames(game_1, frame_idx_1, game_2, frame_idx_2, save_path=None):
+    frame_1 = game_1.tracking_data[game_1.tracking_data["frame"] == frame_idx_1].iloc[0]
+    frame_2 = game_2.tracking_data[game_2.tracking_data["frame"] == frame_idx_2].iloc[0]
+    pitch_control_1 = get_pitch_control_single_frame(
+        frame_1,
+        game_1.pitch_dimensions,
+        n_x_bins=106,
+        n_y_bins=68,
     )
+    pitch_control_2 = get_pitch_control_single_frame(
+        frame_2,
+        game_2.pitch_dimensions,
+        n_x_bins=106,
+        n_y_bins=68,
+    )
+    cmap_red_green = LinearSegmentedColormap.from_list(
+        "reds", [(0, 1, 0, 1), (0.5, 0.5, 0, 0), (1, 0, 0, 1)]
+    )
+
+    fig, ax = plot_soccer_pitch(field_dimen=game.pitch_dimensions, pitch_color="white")
     idx_relative_to_game_1 = game_1.tracking_data[
         game_1.tracking_data["frame"] == frame_idx_1
     ].index[0]
@@ -89,85 +78,31 @@ def diff_frames(game_1, frame_idx_1, game_2, frame_idx_2, save_path=None):
         game_2.tracking_data["frame"] == frame_idx_2
     ].index[0]
 
-    before_alpha = 0.2
-
     fig, ax = plot_tracking_data(
         game_1,
         idx_relative_to_game_1,
         team_colors=["green", "red"],
         ax=ax,
         fig=fig,
-        alpha=before_alpha,
-        # heatmap_overlay=pitch_control_1,
-        # overlay_cmap=cmap_red_green,
-        # add_velocities=False,
+        heatmap_overlay=pitch_control_1,
+        overlay_cmap=cmap_red_green,
+        add_velocities=True,
     )
     fig, ax = plot_tracking_data(
         game_2,
         idx_relative_to_game_2,
-        team_colors=["green", "red"],
+        # RGBA must use 0–1 floats (not 0–255) for matplotlib
+        team_colors=[(0.0, 1.0, 0.0, 0.5), (1.0, 0.0, 0.0, 0.5)],
         ax=ax,
         fig=fig,
-        # heatmap_overlay=pitch_control_2,
-        # overlay_cmap=cmap_red_green,
-        # add_velocities=False,
+        heatmap_overlay=pitch_control_2,
+        overlay_cmap=cmap_red_green,
+        add_velocities=True,
     )
-
-    after_legend_markers = (
-        Line2D(
-            [0],
-            [0],
-            marker="o",
-            color="w",
-            markerfacecolor=(0.0, 1.0, 0.0, 1.0),
-            markeredgecolor="green",
-            markersize=9,
-            linestyle="None",
-        ),
-        Line2D(
-            [0],
-            [0],
-            marker="o",
-            color="w",
-            markerfacecolor=(1.0, 0.0, 0.0, 1.0),
-            markeredgecolor="red",
-            markersize=9,
-            linestyle="None",
-        ),
-    )
-    before_legend_markers = (
-        Line2D(
-            [0],
-            [0],
-            marker="o",
-            color="w",
-            markerfacecolor=(0.0, 1.0, 0.0, before_alpha),
-            markeredgecolor=(0.0, 0.5, 0.0, 0.9),
-            markersize=9,
-            linestyle="None",
-        ),
-        Line2D(
-            [0],
-            [0],
-            marker="o",
-            color="w",
-            markerfacecolor=(1.0, 0.0, 0.0, before_alpha),
-            markeredgecolor=(0.7, 0.0, 0.0, 0.9),
-            markersize=9,
-            linestyle="None",
-        ),
-    )
-    ax.legend(
-        [before_legend_markers, after_legend_markers],
-        ["Before Optimization", "After Optimization"],
-        handler_map={tuple: HandlerTuple(ndivide=None)},
-        loc="lower right",
-        framealpha=0.95,
-    ).set_zorder(20)
-
     if save_path:
         fig.savefig(save_path)
     return fig, ax
+
 
 match_id = 1886347
 tracking_data_github_url = f"https://media.githubusercontent.com/media/SkillCorner/opendata/master/data/matches/{match_id}/{match_id}_tracking_extrapolated.jsonl"
@@ -213,11 +148,10 @@ attacking_team_influence = get_team_influence(
     grid=grid,
     player_ball_distances=None,
 )
-
-result = optimize_tracking_frame(
-    game, 
-    selected_frame_idx, 
-    objective_terms = [
+annealer = SimulatedAnnealing(
+    game,
+    selected_frame_idx,
+    objective_terms=[
         WeightedPitchControlObjective(
             grid=grid,
             attacking_team=attacking_team,
@@ -228,13 +162,19 @@ result = optimize_tracking_frame(
         PressureObjective(
             attacking_player_ids=game.get_column_ids(team=defending_team), game=game
         ),
-    ], 
-    weights = [1,1], 
-    constraints = [TTIConstraint()],
-    algorithm = SimulatedAnnealing,
+    ],
+    constraints=[TTIConstraint()],
+    weights=[1, 1],
+    distance_perturbation=0.2,  # how many yards to randomly move the player on each iteration
+    max_tti=2,  # the maximum time to intercept, i.e. don't move a player if they can't reach that space within 1s
+    num_iterations=1000,  # the number of perturbations to do
 )
+
+result = annealer.run()
 print(result.best_result)
-new_game = deepcopy(game)
-new_game.tracking_data = pd.DataFrame(result.best_frame).transpose().reset_index()
-# generate_fig(new_game, selected_frame_idx, save_path="test.png")
-diff_frames(game, selected_frame_idx, new_game, selected_frame_idx, save_path="test_diff.png")
+# new_game = deepcopy(game)
+# new_game.tracking_data = pd.DataFrame(result.best_frame).transpose().reset_index()
+# # generate_fig(new_game, selected_frame_idx, save_path="test.png")
+# diff_frames(
+#     game, selected_frame_idx, new_game, selected_frame_idx, save_path="test_diff.png"
+# )
